Refine vault functions. Begin refactor.

+++ index.js
@@ -3,33 +3,14 @@
 const { program } = require("commander");
 const fs = require("fs");
 const path = require("path");
+const helpers = require("./helpers");
 
-const settingsJson = path.resolve(__dirname, "settings.json");
+const settingsJson = helpers.settingsJson; // Load vaults config
 
-const vaults = loadPaths();
-const vaultNames = vaults.map((vault) => vault.vaultName);
+const vaults = helpers.loadPaths(); // Vaults array of obj
+const vaultNames = vaults.map((vault) => vault.vaultName); // Vaults array of str names
 
-function loadPaths() {
-    if (fs.existsSync(settingsJson)) {
-        const settingsData = fs.readFileSync(settingsJson);
-        return JSON.parse(settingsData).vaults;
-    }
-    return [];
-}
-
-function saveVault(path, vaultName) {
-    // Check for dupe names
-    for (let entry of vaults) {
-        if (entry.vaultName === vaultName) {
-            console.log(
-                "Error: Vault name already taken. Please provide a unique identifier."
-            );
-            return;
-        }
-    }
-    vaults.push({ path: path, vaultName: vaultName, activeVault: false });
-    fs.writeFileSync(settingsJson, JSON.stringify({ vaults: vaults }, null, 2));
-}
+const saveVault = helpers.saveVault;
 
 program
     .version("1.0.0")
@@ -43,16 +24,19 @@ program
     .command("activate <vaultName>")
     .description("switch active vault to specified vault name")
     .action((vaultName) => {
-        console.log("Setting active vault to:", vaultName);
+        let isVaultExist = false;
         for (let entry of vaults) {
             if (entry.activeVault) {
                 entry.activeVault = false;
             }
             if (entry.vaultName === vaultName) {
+                isVaultExist = true;
                 entry.activeVault = true;
             }
         }
-        fs.writeFileSync(
+        if (!isVaultExist) return console.log("Vault not found, use lawa vault --help for details") // Vault not on file
+        console.log("Active vault set to:", vaultName);
+        fs.writeFileSync( 
             settingsJson,
             JSON.stringify({ vaults: vaults }, null, 2)
         );
@@ -70,6 +54,7 @@ program
     .action((options) => {
         if (!options.add) {
             for (let entry of vaults) {
+                
                 if (entry.activeVault) {
                     if (options.verbose) {
                         console.log(
@@ -78,11 +63,14 @@ program
                             "at path",
                             entry.path
                         );
+                        return
                     } else {
                         console.log("Currently active vault:", entry.vaultName);
+                        return
                     }
                 }
             }
+            console.log("No currently active vault, use lawa activate <vaultName>")
         } else if (options.add) {
             let finalName = "";
             if (options.name) {
@@ -95,7 +83,7 @@ program
                 } else finalName = finalName[finalName.length - 1]; // Write some tests here
             }
             console.log(`Adding vault ${finalName} with path ${options.add}`);
-            saveVault(options.add, finalName);
+            saveVault(options.add, finalName, vaults);
         }
     });
